Componentized articles.
