+++ blog-homepage/src/App.css
@@ -38,3 +38,26 @@
     transform: rotate(360deg);
   }
 }
+
+section {
+  margin-bottom: 100px;
+}
+
+h3 {
+  border-bottom: 1px solid #e9e9e9;
+  font-size: 25px;
+  padding-bottom: 25px;
+}
+
+.articles.grid-2-col {
+  display: grid;
+  grid-gap: 1rem;
+  grid-template-columns: repeat(2, 1fr);
+  
+}
+
+.articles.grid-3-col {
+  display: grid;
+  grid-gap: 1rem;
+  grid-template-columns: repeat(3, 1fr);
+}
